Fix the way flattened_executor projects its base_index_types into index_types
