+++ agency/detail/type_traits.hpp
@@ -1,60 +1,8 @@
 #pragma once
 
 #include <type_traits>
-#include <tuple>
 #include <agency/detail/integer_sequence.hpp>
 #include <agency/detail/type_list.hpp>
-
-#define __DEFINE_HAS_NESTED_TYPE(trait_name, nested_type_name) \
-template<typename T> \
-  struct trait_name  \
-{                    \
-  typedef char yes_type; \
-  typedef int  no_type;  \
-  template<typename S> static yes_type test(typename S::nested_type_name *); \
-  template<typename S> static no_type  test(...); \
-  static bool const value = sizeof(test<T>(0)) == sizeof(yes_type);\
-  typedef std::integral_constant<bool, value> type;\
-};
-
-#define __DEFINE_HAS_NESTED_CLASS_TEMPLATE(trait_name, nested_class_template_name) \
-template<typename T, typename... Types> \
-  struct trait_name         \
-{                           \
-  typedef char yes_type;    \
-  typedef int  no_type;     \
-  template<typename S> static yes_type test(typename S::template nested_class_template_name<Types...> *); \
-  template<typename S> static no_type  test(...); \
-  static bool const value = sizeof(test<T>(0)) == sizeof(yes_type);\
-  typedef std::integral_constant<bool, value> type;\
-};
-
-#ifdef __NVCC__
-#define __DEFINE_HAS_NESTED_MEMBER(trait_name, nested_member_name) \
-template<typename T> \
-  struct trait_name  \
-{                    \
-  typedef char yes_type; \
-  typedef int  no_type;  \
-  template<int i> struct swallow_int {}; \
-  template<typename S> static yes_type test(swallow_int<sizeof(S::nested_member_name)>*); \
-  template<typename S> static no_type  test(...); \
-  static bool const value = sizeof(test<T>(0)) == sizeof(yes_type);\
-  typedef std::integral_constant<bool, value> type;\
-};
-#else
-#define __DEFINE_HAS_NESTED_MEMBER(trait_name, nested_member_name) \
-template<typename T> \
-  struct trait_name  \
-{                    \
-  typedef char yes_type; \
-  typedef int  no_type;  \
-  template<typename S> static yes_type test(decltype(S::nested_member_name)*); \
-  template<typename S> static no_type  test(...); \
-  static bool const value = sizeof(test<T>(0)) == sizeof(yes_type);\
-  typedef std::integral_constant<bool, value> type;\
-};
-#endif
 
 namespace agency
 {
@@ -139,76 +87,6 @@ struct static_or<Condition, Conditions...>
       Condition::value || static_or<Conditions...>::value
     >
 {};
-
-
-__DEFINE_HAS_NESTED_MEMBER(has_value, value);
-
-
-template<class T>
-struct is_tuple : has_value<std::tuple_size<T>> {};
-
-
-template<class Indices, class Tuple>
-struct tuple_type_list_impl;
-
-template<size_t... Indices, class Tuple>
-struct tuple_type_list_impl<index_sequence<Indices...>, Tuple>
-{
-  using type = type_list<
-    typename std::tuple_element<Indices,Tuple>::type...
-  >;
-};
-
-
-template<class T, class Enable = void>
-struct tuple_type_list;
-
-
-template<class Tuple>
-struct tuple_type_list<Tuple, typename std::enable_if<is_tuple<Tuple>::value>::type>
-{
-  using type = typename tuple_type_list_impl<
-    make_index_sequence<std::tuple_size<Tuple>::value>,
-    Tuple
-  >::type;
-};
-
-
-template<class>
-struct is_empty_tuple;
-
-
-template<class T>
-struct is_empty_tuple_impl_impl;
-
-
-template<class... Types>
-struct is_empty_tuple_impl_impl<type_list<Types...>>
-{
-  using type = static_and<
-    static_or<
-      std::is_empty<Types>,
-      is_empty_tuple<Types>
-    >...
-  >;
-};
-
-
-template<class T, class Enable = void>
-struct is_empty_tuple_impl : std::false_type {};
-
-
-template<class Tuple>
-struct is_empty_tuple_impl<Tuple, typename std::enable_if<is_tuple<Tuple>::value>::type>
-{
-  using type = typename is_empty_tuple_impl_impl<
-    typename tuple_type_list<Tuple>::type
-  >::type;
-};
-
-
-template<class Tuple>
-struct is_empty_tuple : is_empty_tuple_impl<Tuple>::type {};
 
 
 template<class T>
