+++ test_executor_array.cu
@@ -49,6 +49,11 @@ int main()
 
     executor_type exec(2);
 
+    for(size_t i = 0; i < 2; ++i)
+    {
+      exec[i].gpu(i);
+    }
+
     auto shape = exec.make_shape(2,{2,2});
 
     auto past = traits::make_ready_future<int>(exec, 13);
@@ -83,6 +88,11 @@ int main()
 
     executor_type exec(2);
 
+    for(size_t i = 0; i < 2; ++i)
+    {
+      exec[i].gpu(i);
+    }
+
     auto shape = exec.make_shape(2,{2,2});
 
     auto f = exec.async_execute([] __device__ (const index_type& idx, int& outer_shared, int& inner_shared, int& inner_inner_shared)
@@ -104,40 +114,46 @@ int main()
     assert(std::all_of(results.begin(), results.end(), [](int x){ return x == 13 + 7 + 42 + 1; }));
   }
 
-  //{
-  //  // test flattened executor_array
-  //  using executor_array_type = executor_array<inner_executor_type, outer_executor_type>;
-  //  using executor_type = flattened_executor<executor_array_type>;
+  {
+    // test flattened executor_array
+    using executor_array_type = executor_array<inner_executor_type, outer_executor_type>;
+    using executor_type = flattened_executor<executor_array_type>;
 
-  //  using traits = agency::executor_traits<executor_type>;
-  //  using shape_type = typename traits::shape_type;
-  //  using index_type = typename traits::index_type;
+    using traits = agency::executor_traits<executor_type>;
+    using shape_type = typename traits::shape_type;
+    using index_type = typename traits::index_type;
 
-  //  executor_array_type exec_array(2);
-  //  executor_type exec{exec_array};
+    executor_array_type exec_array(2);
 
-  //  shape_type shape{2, 2};
+    for(size_t i = 0; i < 2; ++i)
+    {
+      exec_array[i].gpu(i);
+    }
 
-  //  auto ready = traits::make_ready_future<void>(exec);
+    executor_type exec{exec_array};
 
-  //  auto f = exec.then_execute([] __device__ (const index_type& idx, int& outer_shared, int& inner_shared)
-  //  {
-  //    return 13 + outer_shared + inner_shared;
-  //  },
-  //  [](shape_type shape)
-  //  {
-  //    return traits::container<int>(shape);
-  //  },
-  //  shape,
-  //  ready,
-  //  make_7(),
-  //  make_42());
+    shape_type shape{2, 2};
 
-  //  auto results = f.get();
+    auto ready = traits::make_ready_future<void>(exec);
 
-  //  assert(results.size() == agency::detail::shape_size(shape));
-  //  assert(std::all_of(results.begin(), results.end(), [](int x){ return x == 13 + 7; }));
-  //}
+    auto f = exec.then_execute([] __device__ (const index_type& idx, int& outer_shared, int& inner_shared)
+    {
+      return 13 + outer_shared + inner_shared;
+    },
+    [](shape_type shape)
+    {
+      return traits::container<int>(shape);
+    },
+    shape,
+    ready,
+    make_7(),
+    make_42());
+
+    auto results = f.get();
+
+    assert(results.size() == agency::detail::shape_size(shape));
+    assert(std::all_of(results.begin(), results.end(), [](int x){ return x == 13 + 7 + 42; }));
+  }
 
   std::cout << "OK" << std::endl;
